Refactored App/Results/Nav/ColumnHead to avoid passing results down the hierarchy
