+++ src/components/ColumnHead.tsx
@@ -1,20 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import './App.css';
-import { SortableColumns, SortStatus } from '../helpers/enums';
+import { SortStatus } from '../helpers/enums';
 
-function ColumnHead({ name, filteredResults, setFilteredResults, sortedColumn, setSortedColumn }) {
+function ColumnHead({ name, isSorted, setSortedColumn }:{name:string, isSorted: boolean, setSortedColumn:(column:string, reversed:boolean) => void}) {
+  const [ reversed, setReversed ] = useState<boolean>(false);
 
-  const [reversed, setReversed] = useState(true)
+  useEffect(() => {
+    setReversed( isSorted ? false : true)
+  }, [isSorted]);
 
-  const sortFunctions = sortFunctionsArray(name)
 
-  function sortFunctionsArray(name) {
-    if (name === SortableColumns.Amount) { return [dateCompare, stringCompare, amountCompare] }
-    if (name === SortableColumns.Result) { return [amountCompare, dateCompare, stringCompare] }
-    return [amountCompare, stringCompare, dateCompare]
-  }
-
-  function stringCompare(a, b) {
+  function stringCompare(a:any, b:any) {
     let compare = a.orgName.localeCompare(b.orgName)
 
     if (compare === 0) {
@@ -30,17 +26,17 @@ function ColumnHead({ name, filteredResults, setFilteredResults, sortedColumn, s
     return compare
   }
 
-  function dateCompare(a, b) {
+  function dateCompare(a:any, b:any) {
       let yearCompare = Math.sign(a.year - b.year)
       return yearCompare === 0 ? Math.sign(a.month - b.month) : yearCompare
   }
 
-  function amountCompare(a, b) {
+  function amountCompare(a:any, b:any) {
       return Math.sign(a.amount - b.amount)
   }
 
   function renderSortButton() {
-    if (sortedColumn === name) {
+    if (isSorted) {
       return (
         // <button onClick={e => handleSortButton()}>
         <>
@@ -50,21 +46,19 @@ function ColumnHead({ name, filteredResults, setFilteredResults, sortedColumn, s
     }
   }
 
+
   function handleSortButton() {
-    let newResults = filteredResults
-    setReversed(!reversed)
-    for (let i = 0; i < sortFunctions.length; i++) { newResults = newResults.sort(sortFunctions[i]) }
-    if (!reversed) { newResults = newResults.reverse() }
-    setFilteredResults([...newResults])
-    setSortedColumn(name)
+    let rev = !reversed
+    setSortedColumn({column: name, reversed: rev});
+    setReversed(rev)
   }
 
   function activeSortClass() {
-    return (sortedColumn !== name ? 'db__results-sorted-inactive' : 'db__results-sorted-active')  
+    return (isSorted ? 'db__results-sorted-inactive' : 'db__results-sorted-active')  
   }
 
   function sortDirection() {
-    return (sortedColumn !== name 
+    return (!isSorted 
         ? ''
         : reversed ? 'db__results-sorted-ascending' : 'db__results-sorted-descending'
     )
