+++ src/components/Nav.tsx
@@ -1,33 +1,25 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import './App.css';
 import ColumnHead from './ColumnHead';
 import { SortableColumns } from '../helpers/enums';
 
-function Nav({ filteredResults, setFilteredResults }) {
-
-    const [sortedColumn, setSortedColumn] = useState(SortableColumns.ApprovalDate)
+function Nav({ sortedColumn, setSortedColumn }:{ sortedColumn:string, setSortedColumn:(columns:string, reversed:boolean) => void }) {
 
     return (
         <>
             <ColumnHead 
                 name={SortableColumns.Amount} 
-                filteredResults={filteredResults} 
-                setFilteredResults={setFilteredResults}
-                sortedColumn={sortedColumn}
+                isSorted={sortedColumn == SortableColumns.Amount}
                 setSortedColumn={setSortedColumn}
             /> 
             <ColumnHead 
                 name={SortableColumns.Result} 
-                filteredResults={filteredResults} 
-                setFilteredResults={setFilteredResults}
-                sortedColumn={sortedColumn}
+                isSorted={sortedColumn == SortableColumns.Result}
                 setSortedColumn={setSortedColumn}
             />
             <ColumnHead 
                 name={SortableColumns.ApprovalDate} 
-                filteredResults={filteredResults} 
-                setFilteredResults={setFilteredResults}
-                sortedColumn={sortedColumn}
+                isSorted={sortedColumn == SortableColumns.ApprovalDate}
                 setSortedColumn={setSortedColumn}
             />             
         </>
